Refactor: 패키징 변경, 내비게이션 마이그레이션
