+++ app/src/main/java/com/konkuk/hackathon/core/navigation/volunteer/VolunteerNavHost.kt
@@ -4,6 +4,10 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.navigation
+import com.konkuk.hackathon.core.navigation.Route
+import com.konkuk.hackathon.feature.volunteer.home.screen.RecordSubmitScreen
+import com.konkuk.hackathon.feature.volunteer.home.screen.VolunteerHomeScreen
 
 @Composable
 fun VolunteerNavHost(
@@ -18,7 +22,18 @@ fun VolunteerNavHost(
     ) {
         // Home
         composable<VolunteerTabRoute.Home> {
+            VolunteerHomeScreen(
+                padding = padding,
+                navigateToRecordSubmit = { navController.navigate(VolunteerRoute.HomeGraph) })
+        }
 
+        // Home Nested Graph
+        navigation<VolunteerRoute.HomeGraph>(
+            startDestination = VolunteerRoute.RecordSubmit
+        ) {
+            composable<VolunteerRoute.RecordSubmit> {
+                RecordSubmitScreen(padding = padding)
+            }
         }
 
         // Record
