+++ app/src/main/java/com/konkuk/hackathon/core/navigation/volunteer/VolunteerRoute.kt
@@ -3,7 +3,11 @@ package com.konkuk.hackathon.core.navigation.volunteer
 import kotlinx.serialization.Serializable
 
 sealed interface VolunteerRoute {
+    @Serializable
+    data object HomeGraph : VolunteerRoute
 
+    @Serializable
+    data object RecordSubmit : VolunteerRoute
 }
 
 sealed interface VolunteerTabRoute : VolunteerRoute {
